Friendly fire support.

+++ src/AudioShooter/Assets/_Scripts/Controllers/MissileController.cs
@@ -5,16 +5,18 @@ public class MissileController : MonoBehaviour {
 
 	Vector3 _target;
 	Vector3 _direction;
+	bool _friendlyFire;
 
 	public float _velocity;
 	public float _distance;
 
 	public GameObject Shooter { get; private set; }
 
-	public void Initialize(GameObject shooter, Vector3 direction)
+	public void Initialize(GameObject shooter, Vector3 direction, bool friendlyFire)
 	{
 		Shooter = shooter;
 		_direction = direction;
+		_friendlyFire = friendlyFire;
 		_target = transform.position + (_direction * _distance);
 	}
 
@@ -25,7 +27,9 @@ public class MissileController : MonoBehaviour {
 
 	void OnTriggerEnter(Collider other)
 	{
-		if ((other.IsEnemy() && other.gameObject != Shooter) || other.IsMountain())
+		var allowFriendlyFire = _friendlyFire || (this.Shooter.transform.IsEnemy() ^ other.IsEnemy());
+
+		if (allowFriendlyFire && ((other.IsEnemy() && other.gameObject != Shooter)) || other.IsMountain())
 		{
 			MissileAppService.DestroyMissile(gameObject);
 		}
